Add converter implementation
At first I tried to use "roman-numbers" package as it fits the requirements and is covered with tests.
But it doesn't work with jest because of some issue with ramda dependency. So, I had to switch to another library and I added tests anyway.
I checked this library size using bundlephobia.

I understand that the logic of conversion is pretty straightforward, and I can implement it myself, but I decided not to do that for simplicity and speed.
Most of the time we should not spend time on something that was created many times already and which we can reuse unless we have some issues with a third-party solution.